Attempt to use hiedb to resolve source file before resolving using the file tree (#23)
Check hiedb for a src file path before using

Generally hiedb will be more accurate if a source is available there
(but will only be available if the user has used the `src-file-path`
argument when indexing. This will also enable one to index their
dependencies and allow static-ls to jump to them.